+++ src/main/resources/static/js/time.js
@@ -109,7 +109,7 @@ function requestCreate(data) {
 
   return fetch(API_ENDPOINT, requestOptions)
       .then(response => {
-        if (response.status === 200) return response.json();
+        if (response.status === 201) return response.json();
         throw new Error('Create failed');
       });
 }
@@ -129,6 +129,6 @@ function requestDelete(id) {
 
   return fetch(`${API_ENDPOINT}/${id}`, requestOptions)
       .then(response => {
-        if (response.status !== 200) throw new Error('Delete failed');
+        if (response.status !== 204) throw new Error('Delete failed');
       });
 }
